==> 01_city_gps_search.ipynb <==
## STEP 1 - Scrape GPS coordinates

## Search GPS coordinates of all the cities provided by Marketing team from the https://nominatim.org/

Create the file contain the list of the city provided by Marketing Teamand their gps coordinates

import requests
import json
import pandas as pd

Scope of the project: top-35 cities to visit in France

cities = ["Mont Saint Michel",
"St Malo",
"Bayeux",
"Le Havre",
"Rouen",
"Paris",
"Amiens",
"Lille",
"Strasbourg",
"Chateau du Haut Koenigsbourg",
"Colmar",
"Eguisheim",
"Besancon",
"Dijon",
"Annecy",
"Grenoble",
"Lyon",
"Gorges du Verdon",
"Bormes les Mimosas",
"Cassis",
"Marseille",
"Aix en Provence",
"Avignon",
"Uzes",
"Nimes",
"Aigues Mortes",
"Saintes Maries de la mer",
"Collioure",
"Carcassonne",
"Ariege",
"Toulouse",
"Montauban",
"Biarritz",
"Bayonne",
"La Rochelle"]

Create a dict contain the cities and their gps coordinates

citiesGeo = []
for indx,city in enumerate(cities):
    url = "https://nominatim.openstreetmap.org/search?q=%s+Fr&format=json"%(city)
    response = requests.get(url)
    if response.ok:
        cityGeo = {}
        data = json.loads(response.text)
        cityGeo.update({"CityID":indx,
                        "CityName":city,
                        "CityLat":data[0].get("lat"),
                        "CityLon":data[0].get("lon")})
    citiesGeo.append(cityGeo)

result = pd.DataFrame.from_dict(citiesGeo,orient='columns')
result.head()

#Check the empty data
result.isna().sum().sum()

print("Number cities in the initial list :", len(cities))
print("Number cities in the final list :", len(citiesGeo))
assert len(cities)==len(citiesGeo)

Save this dict in the csv file

result.to_csv("refCity.csv",index=False)
