==> 02_search_better_weather_by_city.ipynb <==
## STEP 2
## Search the weather for the city list provided by the marketing team

Use https://openweathermap.org/appid (you have to subscribe to get a free apikey) and https://openweathermap.org/api/one-call-api to get some information about the weather for the 35 cities and put it in a DataFrame

import requests
import pandas as pd
import json
from datetime import datetime, timezone

# api key to connect 
api_key = "YOUR_API_KEY"

#open the city list file
cityGeo = pd.read_csv("refCity.csv")
cityGeo.head()

cityGeo.isna().sum().sum()

columnsToUse = ['pop','feels_like','temp','clouds','wind_speed']
# sorting rule for the selection
sortAscending = [False,False,False,True,True]
dataWeatherFinal = pd.DataFrame(columns = ['From','To','CityID','CityName','CityLat','CityLon'] + columnsToUse)
dataWeatherFinal.head()

Search the weather data to each city

for indx, row in cityGeo.iterrows():
    lat=row["CityLat"]
    lon=row["CityLon"]
    url="https://api.openweathermap.org/data/2.5/onecall?lat=%s&lon=%s&exclude=current,hourly,minutely&appid=%s&units=metric" % (lat, lon, api_key)
    response = requests.get(url)
    data = json.loads(response.text)
    dataWeather = pd.DataFrame(data.get('daily'))
    ##print(dataWeather.head())
    dataWeather['temp'] = dataWeather['temp'].apply(lambda x: x.get('day'))
    dataWeather['feels_like'] = dataWeather['feels_like'].apply(lambda x: x.get('day'))
    dateFrom = pd.to_datetime(datetime.fromtimestamp(dataWeather['dt'].min()).isoformat())
    dateTo = pd.to_datetime(datetime.fromtimestamp(dataWeather['dt'].max()).isoformat())    
    dataToInsert = dataWeather[columnsToUse].mean().to_dict()
    dataToInsert.update({'From':dateFrom,'To':dateTo,'CityName':row["CityName"],'CityID':row["CityID"],'CityLat':lat,'CityLon':lon})
    dataWeatherFinal = dataWeatherFinal.append(dataToInsert,ignore_index=True)

print(dataWeatherFinal.shape)
dataWeatherFinal.head()

#dataWeatherFinal.to_csv("weather_data.csv",index=False)

Search the top 5 cities by best weather

dataWeatherTop5 = dataWeatherFinal.sort_values(by = columnsToUse, ascending = sortAscending).reset_index(drop=True).head(5)

dataWeatherTop5.head()

dataWeatherTop5.rename(columns={"pop":"Pop","feels_like":"FeelLike","temp":"Temp","clouds":"Clouds" ,"wind_speed":"WindSpeed"},inplace=True)

dataWeatherTop5.columns

Save the final data in the csv file

dataWeatherTop5.head(5).to_csv('weather_for_top5_city.csv', index=False)
