==> 03_Booking_scraping.ipynb <==
## STEP 3 -Scrape Booking.com

Data to scrap:
 - hotel name,
 - Url to its booking.com page,
 - Its coordinates: latitude and longitude
 - Score given by the website users
 - Text description of the hotel

import numpy as np
import pandas as pd
import requests
from requests import get
from bs4 import BeautifulSoup
import json

url_base="https://www.booking.com/searchresults.html?ss={}&order=bayesian_review_score"
REQUEST_HEADER = {
                 "User-Agent": "Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/50.0.2661.75 "
                  "Safari/537.36"}

url_base

cityGeo = pd.read_csv("weather_for_top5_city.csv")
cityGeo.head()

cols =['CityID','HotelName','HotelLat', 'HotelLon','HotelURL','HotelScore','HotelDescription']
resultHotels = pd.DataFrame(columns = cols)
for indx, row in cityGeo.iterrows():
    resultHotel={'CityID':row['CityID'],'CityName':row['CityName']}
    request = requests.get(url_base.format(row['CityName']),headers=REQUEST_HEADER)
    soup = BeautifulSoup(request.content, "lxml")
    hotels = soup.find_all("div", {"data-testid": "property-card"})
    for elem in hotels[:20]:
        resultHotel.update({'HotelName':elem.find("div", {"data-testid":"title"}).text})
        hotelUrl =elem.a['href']
        resultHotel.update({'HotelURL':hotelUrl})
        hotelScore = elem.find("div", {"data-testid":"review-score"})
        if hotelScore is not None:
            resultHotel.update({'HotelScore':elem.find("div", {"data-testid":"review-score"}).find("div").text})
        hotelDescriptionTag = elem.find("span", {"data-testid":"distance"})
        if hotelDescriptionTag is not None:
            resultHotel.update({'HotelDescription':elem.find("span", {"data-testid":"distance"}).find_next("div").text})
        details_hotel = BeautifulSoup(requests.get(hotelUrl, headers=REQUEST_HEADER).text,"lxml")
        resultHotel.update({'HotelLat':details_hotel.find("p",{'id':"showMap2"}).a['data-atlas-latlng'].split(',')[0]})
        resultHotel.update({'HotelLon':details_hotel.find("p",{'id':"showMap2"}).a['data-atlas-latlng'].split(',')[1]})
        resultHotels = resultHotels.append(resultHotel,ignore_index=True) 

result = pd.DataFrame.from_dict(resultHotels,orient='columns')
print(result.shape)
result.head()

result.to_csv("top20hotels_weather_data.csv", index=False)
