==> 05_ETL.ipynb <==
## Etape 5 ETL

Create a SQL Database using AWS RDS, Extract the data from S3, Store it in created DB.

import psycopg2
from sqlalchemy import create_engine
import pandas as pd
import boto3
from sqlalchemy.sql import text

### Extract the data from S3

# please remplace by ypur key
key_id = 'YOUR access_key_id'
secret_access_key = 'YOUR secret_access_key'
session = boto3.Session(aws_access_key_id=key_id, 
                        aws_secret_access_key=secret_access_key)
bucket_name = 'buketkayak'
client = session.client("s3")

responseCity = client.get_object(Bucket=bucket_name, Key="refCity.csv") 
cityGeo = pd.read_csv(responseCity.get("Body"))
cityGeo.head()

responseHotels = client.get_object(Bucket=bucket_name, Key="weather_hotels_data.csv") 
weather_hotels = pd.read_csv(responseHotels.get("Body"))
weather_hotels.head()

weather = weather_hotels[['From', 'To', 'CityID', 'CityName', 'CityLat', 'CityLon',
                         'Pop','FeelLike', 'Temp', 'Clouds', 'WindSpeed']].drop_duplicates().reset_index(drop=True).copy()
weather.insert(0, 'WeatherID', weather.index)
weather.head()

hotels = weather_hotels[['HotelName','HotelLat', 'HotelLon','HotelURL',
                         'HotelScore','HotelDescription','CityID']].reset_index(drop=True).copy()
hotels.insert(0, 'HotelID', hotels.index)
hotels.head()

### Insert the data to SQL Database

# CREATE ENGINE
# Replace YOUR_USERNAME, YOUR_PASSWORD, please note that the HOSTNAME below was deleted
#dialect+driver://username:password@host:port/database
engine = create_engine("postgresql+psycopg2://YOUR_USERNAME:YOUR_PASSWORD@kayak-database.ckfbjkm8egwf.eu-west-3.rds.amazonaws.com:5432/postgres", echo=True)

#Check your connection
try:
    with engine.connect() as con:
        engine.execute("SELECT 1")
    print('engine is valid')
except Exception as e:
    print(f'Engine invalid: {str(e)}')

# Create te table refCity
cityGeo.to_sql("refCity", if_exists = 'replace', con = engine, index=False)

# create the table weatherCity
weather.to_sql("weatherCity",if_exists = 'replace', con = engine, index=False)

# create the table hotelsCity
hotels.to_sql("hotelsCity",if_exists = 'replace', con = engine, index=False)

#check the data in the refCity table
statement = text('SELECT * FROM "refCity" ORDER BY "CityName"')
display(pd.read_sql(statement,engine))

#check the data in the weatherCity table
statement = text('SELECT * FROM "weatherCity" ORDER BY "Pop", "Temp" DESC')
display(pd.read_sql(statement,engine))

#check the data in the hotelsCity table
statement = text('SELECT * FROM "hotelsCity" ORDER BY "HotelScore" DESC')
display(pd.read_sql(statement,engine))
